+++ src/Symfony/Component/PropertyInfo/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.3.0
 -----
 
+* Add `ClassListExtractorInterface` which provide a way to get classes for a specific domain
 * Added the ability to extract property type based on its initial value
 
 4.2.0